# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JavaScript learning platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChapterDifficulty = typeof ChapterDifficulty[keyof typeof ChapterDifficulty];


export const ChapterDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Chapter {
  slug: string;
  title: string;
  description: string;
  order: number;
  category: string;
  difficulty: ChapterDifficulty;
  estimatedMinutes: number;
  /** @nullable */
  completed?: boolean | null;
}

export type ChapterDetailDifficulty = typeof ChapterDetailDifficulty[keyof typeof ChapterDetailDifficulty];


export const ChapterDetailDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export type DiagramType = typeof DiagramType[keyof typeof DiagramType];


export const DiagramType = {
  flowchart: 'flowchart',
  scope: 'scope',
  callstack: 'callstack',
  memory: 'memory',
  'event-loop': 'event-loop',
  'prototype-chain': 'prototype-chain',
} as const;

export interface Diagram {
  id: string;
  type: DiagramType;
  title: string;
  /** Mermaid diagram definition or JSON config */
  data: string;
}

export interface ChapterDetail {
  slug: string;
  title: string;
  description: string;
  order: number;
  category: string;
  difficulty: ChapterDetailDifficulty;
  estimatedMinutes: number;
  /** Full markdown content of the chapter */
  content: string;
  diagrams: Diagram[];
  /** @nullable */
  completed?: boolean | null;
  /** @nullable */
  prevSlug?: string | null;
  /** @nullable */
  nextSlug?: string | null;
}

export interface ProgressSummary {
  totalChapters: number;
  completedChapters: number;
  percentComplete: number;
  completedSlugs: string[];
  streakDays: number;
}

export interface ProgressInput {
  completed: boolean;
}

export interface ProgressEntry {
  slug: string;
  completed: boolean;
  /** @nullable */
  completedAt: string | null;
}

export interface ErrorResponse {
  error: string;
}