# lib/api-zod/src/generated/types/chapter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JavaScript learning platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChapterDifficulty } from './chapterDifficulty';

export interface Chapter {
  slug: string;
  title: string;
  description: string;
  order: number;
  category: string;
  difficulty: ChapterDifficulty;
  estimatedMinutes: number;
  /** @nullable */
  completed?: boolean | null;
}