# lib/api-zod/src/generated/types/chapterDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JavaScript learning platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChapterDetailDifficulty } from './chapterDetailDifficulty';
import type { Diagram } from './diagram';

export interface ChapterDetail {
  slug: string;
  title: string;
  description: string;
  order: number;
  category: string;
  difficulty: ChapterDetailDifficulty;
  estimatedMinutes: number;
  /** Full markdown content of the chapter */
  content: string;
  diagrams: Diagram[];
  /** @nullable */
  completed?: boolean | null;
  /** @nullable */
  prevSlug?: string | null;
  /** @nullable */
  nextSlug?: string | null;
}