# lib/api-zod/src/generated/types/diagram.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JavaScript learning platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DiagramType } from './diagramType';

export interface Diagram {
  id: string;
  type: DiagramType;
  title: string;
  /** Mermaid diagram definition or JSON config */
  data: string;
}