# lib/api-zod/src/generated/types/progressEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JavaScript learning platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressEntry {
  slug: string;
  completed: boolean;
  /** @nullable */
  completedAt: string | null;
}