# lib/api-zod/src/generated/types/progressSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JavaScript learning platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressSummary {
  totalChapters: number;
  completedChapters: number;
  percentComplete: number;
  completedSlugs: string[];
  streakDays: number;
}